% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Component.R
\name{Component_select}
\alias{Component_select}
\title{查询Component}
\usage{
Component_select(erp_token)
}
\arguments{
\item{erp_token}{}
}
\value{
无返回值
}
\description{
查询Component
}
\examples{
Component_select()
}
