% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalInvoice.R
\name{SalInvoice_Compute}
\alias{SalInvoice_Compute}
\title{计算毛利表}
\usage{
SalInvoice_Compute(erp_token, FDate)
}
\arguments{
\item{erp_token}{}

\item{FDate}{}
}
\value{
无返回值
}
\description{
计算毛利表
}
\examples{
SalInvoice_Compute()
}
