% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalInvoice.R
\name{SalInvoice_delete}
\alias{SalInvoice_delete}
\title{删除SalInvoice}
\usage{
SalInvoice_delete(erp_token, FInvoice)
}
\arguments{
\item{erp_token}{}

\item{FInvoice}{}
}
\value{
无返回值
}
\description{
删除SalInvoice
}
\examples{
SalInvoice_delete()
}
