% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalInvoice.R
\name{SalInvoice_select_sum}
\alias{SalInvoice_select_sum}
\title{查询SalInvoice——sum}
\usage{
SalInvoice_select_sum(erp_token, FInvoice, FStartDate, FEndDate)
}
\arguments{
\item{FInvoice}{}

\item{token}{}
}
\value{
无返回值
}
\description{
查询SalInvoice——sum
}
\examples{
SalInvoice_select_sum()
}
